Another iPhone bug - table on Accolades page should probably look like the Accolades page on current site on iPhone. (done, needs push)
Viewpoints index page - probably more the More Press Releases > AFTER the press releases are shown? If that doesn’t make sense, I can show you. Hrm. I get it, but it’s part of the header, so I’ll have to put it in twice and hide on larger screens?

Resources index page - not loving how some of the images in the boxes are getting cropped on phone.